[x2cpg] Decompose AstCreatorBase into focused internal traits
AstCreatorBase had grown to 539 lines with 41 methods spanning five unrelated concerns. Split into four private[x2cpg] traits under internal/, each with a single responsibility:

  - AnnotationAstBuilder: annotationAst, annotationAssignmentAst
  - CallAstBuilder: callAst, fieldAccessAst
  - MethodAstBuilder: methodAst, methodStubAst, staticInitMethodAst, returnAst
  - ControlStructureAstBuilder: while/do/for/if/switch/try/break/continue

AstCreatorBase retains only primitives (blockAst, wrapMultipleInBlock, setArgumentIndices) and mixes the traits in via `with` clauses. The public API is unchanged.

Also removes the unused deprecated `tryCatchAstWithOrder` method, and cleans up redundant `val` modifiers on implicit parameters across affected frontends.
