+++ joern-cli/frontends/x2cpg/src/main/scala/io/joern/x2cpg/AstCreatorBase.scala
@@ -8,8 +8,13 @@ import io.shiftleft.semanticcpg.language.types.structure.NamespaceTraversal
 
 import java.nio.file.Paths
 
-abstract class AstCreatorBase[Node, NodeProcessor](filename: String)(implicit withSchemaValidation: ValidationMode)
-    extends AstNodeBuilder[Node, NodeProcessor] { this: NodeProcessor =>
+abstract class AstCreatorBase[Node, NodeProcessor](filename: String)(implicit val withSchemaValidation: ValidationMode)
+    extends AstNodeBuilder[Node, NodeProcessor]
+    with internal.AnnotationAstBuilder[Node, NodeProcessor]
+    with internal.CallAstBuilder[Node, NodeProcessor]
+    with internal.MethodAstBuilder[Node, NodeProcessor]
+    with internal.ControlStructureAstBuilder[Node, NodeProcessor] { this: NodeProcessor =>
+
   val diffGraph: DiffGraphBuilder = Cpg.newDiffGraphBuilder
 
   private val closureKeyPool = new IntervalKeyPool(first = 0, last = Long.MaxValue)
@@ -28,121 +33,6 @@ abstract class AstCreatorBase[Node, NodeProcessor](filename: String)(implicit wi
       .order(1)
   }
 
-  /** Creates an AST that represents an annotation, including its content (annotation parameter assignments).
-    */
-  def annotationAst(annotation: NewAnnotation, children: Seq[Ast]): Ast = {
-    val annotationAst = Ast(annotation)
-    annotationAst.withChildren(children)
-  }
-
-  /** Creates an AST that represents an annotation assignment with a name for the assigned value, its overall code, and
-    * the respective assignment AST.
-    */
-  def annotationAssignmentAst(assignmentValueName: String, code: String, assignmentAst: Ast): Ast = {
-    val parameter      = NewAnnotationParameter().code(assignmentValueName)
-    val assign         = NewAnnotationParameterAssign().code(code)
-    val assignChildren = List(Ast(parameter), assignmentAst)
-    setArgumentIndices(assignChildren)
-    Ast(assign)
-      .withChild(Ast(parameter))
-      .withChild(assignmentAst)
-  }
-
-  /** Creates an AST that represents an entire method, including its content.
-    */
-  def methodAst(
-    method: NewMethod,
-    parameters: Seq[Ast],
-    body: Ast,
-    methodReturn: NewMethodReturn,
-    modifiers: Seq[NewModifier] = Nil
-  ): Ast =
-    methodAstWithAnnotations(method, parameters, body, methodReturn, modifiers, annotations = Nil)
-
-  /** Creates an AST that represents an entire method, including its content and with support for both method and
-    * parameter annotations.
-    */
-  def methodAstWithAnnotations(
-    method: NewMethod,
-    parameters: Seq[Ast],
-    body: Ast,
-    methodReturn: NewMethodReturn,
-    modifiers: Seq[NewModifier] = Nil,
-    annotations: Seq[Ast] = Nil
-  ): Ast =
-    Ast(method)
-      .withChildren(parameters)
-      .withChild(body)
-      .withChildren(modifiers.map(Ast(_)))
-      .withChildren(annotations)
-      .withChild(Ast(methodReturn))
-
-  /** Creates an AST that represents a method stub, containing information about the method, its parameters, and the
-    * return type.
-    */
-  def methodStubAst(
-    method: NewMethod,
-    parameters: Seq[Ast],
-    methodReturn: NewMethodReturn,
-    modifiers: Seq[NewModifier] = Nil
-  ): Ast =
-    Ast(method)
-      .withChildren(parameters)
-      .withChild(Ast(NewBlock().typeFullName(Defines.Any)))
-      .withChildren(modifiers.map(Ast(_)))
-      .withChild(Ast(methodReturn))
-
-  def staticInitMethodAst(
-    node: Node,
-    initAsts: List[Ast],
-    fullName: String,
-    signature: Option[String],
-    returnType: String,
-    fileName: Option[String] = None
-  ): Ast = {
-    val methodNode_ = methodNode(
-      node,
-      Defines.StaticInitMethodName,
-      fullName,
-      signature.getOrElse(PropertyDefaults.Signature),
-      fileName.getOrElse(PropertyDefaults.Filename)
-    )
-    val staticModifier = NewModifier().modifierType(ModifierTypes.STATIC)
-    val body           = blockAst(NewBlock().typeFullName(Defines.Any), initAsts)
-    val methodReturn   = methodReturnNode(node, returnType)
-    methodAst(methodNode_, Nil, body, methodReturn, List(staticModifier))
-  }
-
-  /** For a given return node and arguments, create an AST that represents the return instruction. The main purpose of
-    * this method is to automatically assign the correct argument indices.
-    */
-  def returnAst(returnNode: NewReturn, arguments: Seq[Ast] = List()): Ast = {
-    setArgumentIndices(arguments)
-    Ast(returnNode)
-      .withChildren(arguments)
-      .withArgEdges(returnNode, arguments.flatMap(_.root))
-  }
-
-  /** For a given node, condition AST and children ASTs, create an AST that represents the control structure. The main
-    * purpose of this method is to automatically assign the correct condition edges.
-    */
-  def controlStructureAst(
-    controlStructureNode: NewControlStructure,
-    condition: Option[Ast],
-    children: Seq[Ast] = Seq(),
-    placeConditionLast: Boolean = false
-  ): Ast = {
-    condition match {
-      case Some(conditionAst) =>
-        Ast(controlStructureNode)
-          .withChildren(if (placeConditionLast) children :+ conditionAst else conditionAst +: children)
-          .withConditionEdges(controlStructureNode, List(conditionAst.root).flatten)
-      case _ =>
-        Ast(controlStructureNode)
-          .withChildren(children)
-    }
-  }
-
   def wrapMultipleInBlock(asts: Seq[Ast], lineNumber: Option[Int]): Ast = {
     asts.toList match {
       case Nil        => blockAst(NewBlock().typeFullName(Defines.Any).lineNumber(lineNumber))
@@ -151,311 +41,11 @@ abstract class AstCreatorBase[Node, NodeProcessor](filename: String)(implicit wi
     }
   }
 
-  def whileAst(
-    condition: Option[Ast],
-    body: Seq[Ast],
-    code: Option[String] = None,
-    lineNumber: Option[Int] = None,
-    columnNumber: Option[Int] = None
-  ): Ast = {
-    var whileNode = NewControlStructure()
-      .controlStructureType(ControlStructureTypes.WHILE)
-      .lineNumber(lineNumber)
-      .columnNumber(columnNumber)
-    if (code.isDefined) {
-      whileNode = whileNode.code(code.get)
-    }
-    val astWithChildren = controlStructureAst(whileNode, condition, body)
-    body.headOption.flatMap(_.root) match {
-      case Some(bodyRoot) => astWithChildren.withTrueBodyEdge(whileNode, bodyRoot)
-      case None           => astWithChildren
-    }
-  }
-
-  def doWhileAst(
-    condition: Option[Ast],
-    body: Seq[Ast],
-    code: Option[String] = None,
-    lineNumber: Option[Int] = None,
-    columnNumber: Option[Int] = None
-  ): Ast = {
-    var doWhileNode = NewControlStructure()
-      .controlStructureType(ControlStructureTypes.DO)
-      .lineNumber(lineNumber)
-      .columnNumber(columnNumber)
-    if (code.isDefined) {
-      doWhileNode = doWhileNode.code(code.get)
-    }
-    val astWithChildren = controlStructureAst(doWhileNode, condition, body, placeConditionLast = true)
-    body.headOption.flatMap(_.root) match {
-      case Some(doBodyRoot) => astWithChildren.withDoBodyEdge(doWhileNode, doBodyRoot)
-      case None             => astWithChildren
-    }
-  }
-
-  def switchAst(switchNode: NewControlStructure, condition: Ast, body: Seq[Ast]): Ast = {
-    val astWithChildren = controlStructureAst(switchNode, Option(condition), body)
-    body.headOption.flatMap(_.root) match {
-      case Some(bodyRoot) => astWithChildren.withTrueBodyEdge(switchNode, bodyRoot)
-      case None           => astWithChildren
-    }
-  }
-
-  def forAst(
-    forNode: NewControlStructure,
-    locals: Seq[Ast],
-    initAsts: Seq[Ast],
-    conditionAsts: Seq[Ast],
-    updateAsts: Seq[Ast],
-    bodyAst: Ast
-  ): Ast =
-    forAst(forNode, locals, initAsts, conditionAsts, updateAsts, Seq(bodyAst))
-
-  private def setOrderExplicitly(ast: Ast, order: Int): Ast = {
-    ast.root match {
-      case Some(value: ExpressionNew) => value.order(order); ast
-      case _                          => ast
-    }
-  }
-
-  def forAst(
-    forNode: NewControlStructure,
-    locals: Seq[Ast],
-    initAsts: Seq[Ast],
-    conditionAsts: Seq[Ast],
-    updateAsts: Seq[Ast],
-    bodyAsts: Seq[Ast]
-  ): Ast = {
-    val lineNumber     = forNode.lineNumber
-    val numOfLocals    = locals.size
-    val initBlock      = setOrderExplicitly(wrapMultipleInBlock(initAsts, lineNumber), numOfLocals + 1)
-    val conditionBlock = setOrderExplicitly(wrapMultipleInBlock(conditionAsts, lineNumber), numOfLocals + 2)
-    val updateBlock    = setOrderExplicitly(wrapMultipleInBlock(updateAsts, lineNumber), numOfLocals + 3)
-    // for the expected orders see CfgCreator.cfgForForStatement
-    if (bodyAsts.nonEmpty) setOrderExplicitly(bodyAsts.head, numOfLocals + 4)
-    val astWithChildren = Ast(forNode)
-      .withChildren(locals)
-      .withChild(initBlock)
-      .withChild(conditionBlock)
-      .withChild(updateBlock)
-      .withChildren(bodyAsts)
-      .withConditionEdges(forNode, conditionAsts.flatMap(_.root).toList)
-
-    val astWithForInit = initBlock.root match {
-      case Some(initRoot) => astWithChildren.withForInitEdge(forNode, initRoot)
-      case None           => astWithChildren
-    }
-
-    val astWithForUpdate = updateBlock.root match {
-      case Some(updateRoot) => astWithForInit.withForUpdateEdge(forNode, updateRoot)
-      case None             => astWithForInit
-    }
-
-    bodyAsts.headOption.flatMap(_.root) match {
-      case Some(bodyRoot) => astWithForUpdate.withForBodyEdge(forNode, bodyRoot)
-      case None           => astWithForUpdate
-    }
-  }
-
-  /** Creates an AST for a break statement. When `labelName` is present a `JumpLabel` child is created at order 1 and
-    * connected to the break node via a `JUMP_ARGUMENT` edge.
-    */
-  def breakAst(node: Node, codeStr: String, labelName: Option[String]): Ast =
-    labeledJumpAst(node, ControlStructureTypes.BREAK, codeStr, labelName)
-
-  /** Creates an AST for a break statement. When `labelNumber` is present a `Literal` child is created at order 1 and
-    * connected to the break node via a `JUMP_ARGUMENT` edge.
-    */
-  def breakAst(node: Node, codeStr: String, labelNumber: Option[Int])(implicit dummy: DummyImplicit): Ast =
-    integerJumpAst(node, ControlStructureTypes.BREAK, codeStr, labelNumber)
-
-  /** Creates an AST for a break statement without a jump argument. */
-  def breakAst(node: Node, codeStr: String): Ast =
-    jumpAst(node, ControlStructureTypes.BREAK, codeStr)
-
-  /** Creates an AST for a continue statement. When `labelName` is present a `JumpLabel` child is created at order 1 and
-    * connected to the continue node via a `JUMP_ARGUMENT` edge.
-    */
-  def continueAst(node: Node, codeStr: String, labelName: Option[String]): Ast =
-    labeledJumpAst(node, ControlStructureTypes.CONTINUE, codeStr, labelName)
-
-  /** Creates an AST for a continue statement. When `labelNumber` is present a `Literal` child is created at order 1 and
-    * connected to the continue node via a `JUMP_ARGUMENT` edge.
-    */
-  def continueAst(node: Node, codeStr: String, labelNumber: Option[Int])(implicit dummy: DummyImplicit): Ast =
-    integerJumpAst(node, ControlStructureTypes.CONTINUE, codeStr, labelNumber)
-
-  /** Creates an AST for a continue statement without a jump argument. */
-  def continueAst(node: Node, codeStr: String): Ast =
-    jumpAst(node, ControlStructureTypes.CONTINUE, codeStr)
-
-  def labeledJumpAst(node: Node, jumpType: String, codeStr: String, labelNumber: Option[Int])(implicit
-    dummy: DummyImplicit
-  ): Ast =
-    integerJumpAst(node, jumpType, codeStr, labelNumber)
-
-  def labeledJumpAst(node: Node, jumpType: String, codeStr: String): Ast =
-    jumpAst(node, jumpType, codeStr)
-
-  def labeledJumpAst(node: Node, jumpType: String, codeStr: String, labelName: Option[String]): Ast = {
-    val jumpNode = NewControlStructure()
-      .parserTypeName(node.getClass.getSimpleName)
-      .controlStructureType(jumpType)
-      .code(codeStr)
-      .lineNumber(line(node))
-      .columnNumber(column(node))
-    labelName match {
-      case Some(name) =>
-        val jumpLabelNode = NewJumpLabel()
-          .parserTypeName(node.getClass.getSimpleName)
-          .name(name)
-          .code(name)
-          .lineNumber(line(node))
-          .columnNumber(column(node))
-          .order(1)
-        Ast(jumpNode)
-          .withChild(Ast(jumpLabelNode))
-          .withJumpArgumentEdge(jumpNode, jumpLabelNode)
-      case None =>
-        Ast(jumpNode)
-    }
-  }
-
-  def integerJumpAst(node: Node, jumpType: String, codeStr: String, labelNumber: Option[Int]): Ast = {
-    val jumpNode = NewControlStructure()
-      .parserTypeName(node.getClass.getSimpleName)
-      .controlStructureType(jumpType)
-      .code(codeStr)
-      .lineNumber(line(node))
-      .columnNumber(column(node))
-    labelNumber match {
-      case Some(number) =>
-        val integerJump = NewLiteral()
-          .code(number.toString)
-          .typeFullName(Defines.Any)
-          .lineNumber(line(node))
-          .columnNumber(column(node))
-          .order(1)
-        Ast(jumpNode)
-          .withChild(Ast(integerJump))
-          .withJumpArgumentEdge(jumpNode, integerJump)
-      case None =>
-        Ast(jumpNode)
-    }
-  }
-
-  private def jumpAst(node: Node, jumpType: String, codeStr: String): Ast =
-    Ast(
-      NewControlStructure()
-        .parserTypeName(node.getClass.getSimpleName)
-        .controlStructureType(jumpType)
-        .code(codeStr)
-        .lineNumber(line(node))
-        .columnNumber(column(node))
-    )
-
-  /** For the given try body, catch ASTs and finally AST, create a try-catch-finally AST with orders set correctly for
-    * the ossdataflow engine.
-    */
-  @deprecated(
-    "This will be removed once all frontends switched to `tryCatchAst` using ControlStructure nodes for catches/finally. Use `tryCatchAst` instead."
-  )
-  def tryCatchAstWithOrder(
-    tryNode: NewControlStructure,
-    tryBodyAst: Ast,
-    catchAsts: Seq[Ast],
-    finallyAst: Option[Ast]
-  ): Ast = {
-    tryBodyAst.root.collect { case x: ExpressionNew => x }.foreach(_.order = 1)
-    catchAsts.flatMap(_.root).collect { case x: ExpressionNew => x }.foreach(_.order = 2)
-    finallyAst.flatMap(_.root).collect { case x: ExpressionNew => x }.foreach(_.order = 3)
-    Ast(tryNode)
-      .withChild(tryBodyAst)
-      .withChildren(catchAsts)
-      .withChildren(finallyAst.toList)
-  }
-
-  /** For the given try body, catch ASTs, and finally AST, create a try-catch-finally AST.
-    */
-  def tryCatchAst(tryNode: NewControlStructure, tryBodyAst: Ast, catchAsts: Seq[Ast], finallyAst: Option[Ast]): Ast = {
-    setArgumentIndices(tryBodyAst +: (catchAsts ++ finallyAst.toSeq))
-    val astWithChildren = Ast(tryNode)
-      .withChild(tryBodyAst)
-      .withChildren(catchAsts)
-      .withChildren(finallyAst.toSeq)
-
-    val astWithTryBody = tryBodyAst.root match {
-      case Some(tryBodyRoot) => astWithChildren.withTryBodyEdge(tryNode, tryBodyRoot)
-      case None              => astWithChildren
-    }
-
-    val astWithCatchBodies = astWithTryBody.withCatchBodyEdges(tryNode, catchAsts.flatMap(_.root).toList)
-
-    finallyAst.flatMap(_.root) match {
-      case Some(finallyRoot) => astWithCatchBodies.withFinallyBodyEdge(tryNode, finallyRoot)
-      case None              => astWithCatchBodies
-    }
-  }
-
   /** For a given block node and statement ASTs, create an AST that represents the block. The main purpose of this
     * method is to increase the readability of the code which creates block asts.
     */
   def blockAst(blockNode: NewBlock, statements: List[Ast] = List()): Ast = {
     Ast(blockNode).withChildren(statements)
-  }
-
-  /** Create an abstract syntax tree for a call, including CPG-specific edges required for arguments and the receiver.
-    *
-    * Our call representation is inspired by ECMAScript, that is, in addition to arguments, a call has a base and a
-    * receiver. For languages other than Javascript, leave `receiver` empty for now.
-    *
-    * @param callNode
-    *   the node that represents the entire call
-    * @param arguments
-    *   arguments (without the base argument (instance))
-    * @param base
-    *   the value to use as `this` in the method call.
-    * @param receiver
-    *   the object in which the property lookup is performed
-    */
-  def callAst(
-    callNode: NewCall,
-    arguments: Seq[Ast] = List(),
-    base: Option[Ast] = None,
-    receiver: Option[Ast] = None
-  ): Ast = {
-
-    setArgumentIndices(arguments)
-
-    val baseRoot = base.flatMap(_.root).toList
-    val bse      = base.getOrElse(Ast())
-    baseRoot match {
-      case List(x: ExpressionNew) =>
-        x.argumentIndex = 0
-      case _ =>
-    }
-
-    val receiverRoot = if (receiver.isEmpty && base.nonEmpty) {
-      baseRoot
-    } else {
-      val r = receiver.flatMap(_.root).toList
-      r match {
-        case List(x: ExpressionNew) =>
-          x.argumentIndex = -1
-        case _ =>
-      }
-      r
-    }
-
-    val rcvAst = receiver.getOrElse(Ast())
-
-    Ast(callNode)
-      .withChild(rcvAst)
-      .withChild(bse)
-      .withChildren(arguments)
-      .withArgEdges(callNode, baseRoot)
-      .withArgEdges(callNode, arguments.flatMap(_.root))
-      .withReceiverEdges(callNode, receiverRoot)
   }
 
   def setArgumentIndices(arguments: Seq[Ast], start: Int = 1): Unit = {
@@ -467,49 +57,14 @@ abstract class AstCreatorBase[Node, NodeProcessor](filename: String)(implicit wi
     }
   }
 
-  def fieldAccessAst(
-    originNode: Node,
-    fieldIdentifierOrigin: Node,
-    base: Ast,
-    code: String,
-    fieldName: String,
-    fieldTypeFullName: String
-  ): Ast = {
-    val callNode_ = callNode(
-      originNode,
-      code,
-      Operators.fieldAccess,
-      Operators.fieldAccess,
-      DispatchTypes.STATIC_DISPATCH,
-      None,
-      Some(fieldTypeFullName)
-    )
-    val fieldIdentifierNode_ = fieldIdentifierNode(fieldIdentifierOrigin, fieldName, fieldName)
-    callAst(callNode_, Seq(base, Ast(fieldIdentifierNode_)))
-  }
-
-  def ifThenElseAst(ifNode: NewControlStructure, conditionAst: Option[Ast], thenAst: Ast, elseAst: Option[Ast]): Ast = {
-    val astWithChildren = controlStructureAst(ifNode, conditionAst, thenAst :: elseAst.toList)
-    val astWithTrueBody = thenAst.root match {
-      case Some(thenRoot) => astWithChildren.withTrueBodyEdge(ifNode, thenRoot)
-      case None           => astWithChildren
-    }
-    val astWithBodies = elseAst.flatMap(_.root) match {
-      case Some(elseRoot) => astWithTrueBody.withFalseBodyEdge(ifNode, elseRoot)
-      case None           => astWithTrueBody
-    }
-
-    astWithBodies
-  }
-
-  def withIndex[T, X](nodes: Seq[T])(f: (T, Int) => X): Seq[X] =
+  def withIndex[T, X](nodes: Seq[T])(func: (T, Int) => X): Seq[X] =
     nodes.zipWithIndex.map { case (x, i) =>
-      f(x, i + 1)
+      func(x, i + 1)
     }
 
-  def withIndex[T, X](nodes: Array[T])(f: (T, Int) => X): Seq[X] =
+  def withIndex[T, X](nodes: Array[T])(func: (T, Int) => X): Seq[X] =
     nodes.toIndexedSeq.zipWithIndex.map { case (x, i) =>
-      f(x, i + 1)
+      func(x, i + 1)
     }
 
   def withArgumentIndex[T <: ExpressionNew](node: T, argIdxOpt: Option[Int]): T = {
